+++ lib/src/api_guide_implementation.dart
@@ -15,23 +15,59 @@ class APIGuide implements APIGuideInterface {
 
   /// Implement the interface
   @override
+
+  /// display Widget implementation
   Widget display({
+    /// Build context implementation
     required BuildContext context,
+
+    /// Host URL implementation
     required String urlHost,
+
+    /// List of API items implementation
     required List<APIItem> apiItems,
+
+    /// List of FAQs implementation
     required List<APIGuideFAQ> apiFaqs,
+
+    /// API version implementation
     required double version,
+
+    /// Date of the latest update implementation
     required DateTime latestUpdate,
+
+    /// Introduction to the API Guide implementation
     required String apiIntro,
   }) {
-    return APIGuideScreen(
-      context: context,
-      urlHost: urlHost,
-      apiItems: apiItems,
-      apiFaqs: apiFaqs,
-      version: version,
-      latestUpdate: latestUpdate,
-      apiIntro: apiIntro,
+    /// Define [ThemeStateNotifier] instance
+    final ThemeStateNotifier themeStateNotifier = ThemeStateNotifier();
+
+    /// Return ThemeNotifierProvider
+    return ThemeNotifierProvider(
+      /// Initialize the themeStateNotifier attribute
+      themeStateNotifier: themeStateNotifier,
+      child: APIGuideScreen(
+        /// Build context implementation
+        context: context,
+
+        /// Host URL implementation
+        urlHost: urlHost,
+
+        /// List of API items implementation
+        apiItems: apiItems,
+
+        /// List of FAQs implementation
+        apiFaqs: apiFaqs,
+
+        /// API version implementation
+        version: version,
+
+        /// Date of the latest update implementation
+        latestUpdate: latestUpdate,
+
+        /// Introduction to the API Guide implementation
+        apiIntro: apiIntro,
+      ),
     );
   }
 }