+++ lib/src/api_guide_interface.dart
@@ -13,26 +13,27 @@ import '../api_guide.dart';
 
 /// Make it as interface to implement later
 abstract class APIGuideInterface {
+  /// display Widget interface to implement later
   Widget display({
-    ///
+    /// Build context interface to implement later
     required BuildContext context,
 
-    ///
+    /// Host URL interface to implement later
     required String urlHost,
 
-    ///
+    /// List of API items interface to implement later
     required List<APIItem> apiItems,
 
-    ///
+    /// List of FAQs interface to implement later
     required List<APIGuideFAQ> apiFaqs,
 
-    ///
+    /// API version interface to implement later
     required double version,
 
-    ///
+    /// Date of the latest update interface to implement later
     required DateTime latestUpdate,
 
-    ///
+    /// Introduction to the API Guide interface to implement later
     required String apiIntro,
   });
 }