+++ lib/src/presentation/presentation.dart
@@ -4,6 +4,7 @@
 
 /// Code starts here
 export '../../src/presentation/helpers/constants.dart';
+export '../../src/presentation/widgets/api_guide_app_bar.dart';
 export '../../src/presentation/widgets/api_guide_screen.dart';
 export '../../src/presentation/widgets/api_item.dart';
 export '../../src/presentation/widgets/api_item_body_end.dart';
@@ -12,7 +13,6 @@ export '../../src/presentation/widgets/api_item_header.dart';
 export '../../src/presentation/widgets/api_item_sample_code.dart';
 export '../../src/presentation/widgets/api_navigator.dart';
 export '../../src/presentation/widgets/api_navigator_drawer.dart';
-export '../../src/presentation/widgets/app_bar.dart';
 export '../../src/presentation/widgets/faq_item.dart';
 
 /// End of code