+++ lib/src/presentation/widgets/api_navigator.dart
@@ -104,6 +104,9 @@ class APIGuideNavigator extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    /// ThemeNotifierProvider to check theme attributes state
+    final themeState = ThemeNotifierProvider.of(context).themeStateNotifier;
+
     return Padding(
       padding: const EdgeInsets.fromLTRB(
         Constants.size0,
@@ -119,7 +122,10 @@ class APIGuideNavigator extends StatelessWidget {
           height: MediaQuery.of(context).size.height,
           child: DecoratedBox(
             decoration: BoxDecoration(
-              color: Constants.greyLightColor,
+              /// Check the current light/dark theme mode
+              color: themeState.isDarkMode
+                  ? ConstantsDarkMode.greyLightColor
+                  : ConstantsLightMode.greyLightColor,
             ),
             child: SingleChildScrollView(
               child: Padding(
@@ -135,7 +141,11 @@ class APIGuideNavigator extends StatelessWidget {
                           '${Constants.apiGuideTxt}$version',
                           style: TextStyle(
                             fontWeight: FontWeight.bold,
-                            color: Constants.blackColor,
+
+                            /// Check the current light/dark theme mode
+                            color: themeState.isDarkMode
+                                ? ConstantsDarkMode.blackColor
+                                : ConstantsLightMode.blackColor,
                             fontSize: Constants.size22,
                           ),
                         )
@@ -150,7 +160,11 @@ class APIGuideNavigator extends StatelessWidget {
                           '${Constants.latestUpdateTxt}${latestUpdate.year}-${latestUpdate.month}-${latestUpdate.day}',
                           style: TextStyle(
                             fontWeight: FontWeight.bold,
-                            color: Constants.blackColor,
+
+                            /// Check the current light/dark theme mode
+                            color: themeState.isDarkMode
+                                ? ConstantsDarkMode.blackColor
+                                : ConstantsLightMode.blackColor,
                             fontSize: Constants.size10,
                           ),
                         )
@@ -165,14 +179,22 @@ class APIGuideNavigator extends StatelessWidget {
                         children: [
                           Icon(
                             Constants.arrowDoubleNavIcon,
-                            color: Constants.indigoColor,
+
+                            /// Check the current light/dark theme mode
+                            color: themeState.isDarkMode
+                                ? ConstantsDarkMode.indigoColor
+                                : ConstantsLightMode.indigoColor,
                             size: Constants.size15,
                           ),
                           Text(
                             Constants.introTxt,
                             style: TextStyle(
                               fontWeight: FontWeight.bold,
-                              color: Constants.indigoColor,
+
+                              /// Check the current light/dark theme mode
+                              color: themeState.isDarkMode
+                                  ? ConstantsDarkMode.indigoColor
+                                  : ConstantsLightMode.indigoColor,
                               fontSize: Constants.size15,
                             ),
                           )
@@ -201,14 +223,26 @@ class APIGuideNavigator extends StatelessWidget {
                                           children: [
                                             Icon(
                                               Constants.arrowDoubleNavIcon,
-                                              color: Constants.indigoColor,
+
+                                              /// Check the current light/dark theme mode
+                                              color: themeState.isDarkMode
+                                                  ? ConstantsDarkMode
+                                                      .indigoColor
+                                                  : ConstantsLightMode
+                                                      .indigoColor,
                                               size: Constants.size15,
                                             ),
                                             Text(
                                               item.title,
                                               style: TextStyle(
                                                 fontWeight: FontWeight.bold,
-                                                color: Constants.indigoColor,
+
+                                                /// Check the current light/dark theme mode
+                                                color: themeState.isDarkMode
+                                                    ? ConstantsDarkMode
+                                                        .indigoColor
+                                                    : ConstantsLightMode
+                                                        .indigoColor,
                                                 fontSize: Constants.size15,
                                               ),
                                             ),
@@ -237,7 +271,12 @@ class APIGuideNavigator extends StatelessWidget {
                                                 child: Text(
                                                   item.request.method.name,
                                                   style: TextStyle(
-                                                    color: Constants.whiteColor,
+                                                    /// Check the current light/dark theme mode
+                                                    color: themeState.isDarkMode
+                                                        ? ConstantsDarkMode
+                                                            .whiteColor
+                                                        : ConstantsLightMode
+                                                            .whiteColor,
                                                     fontWeight: FontWeight.bold,
                                                     fontSize: Constants.size10,
                                                   ),
@@ -262,14 +301,22 @@ class APIGuideNavigator extends StatelessWidget {
                         children: [
                           Icon(
                             Constants.arrowDoubleNavIcon,
-                            color: Constants.indigoColor,
+
+                            /// Check the current light/dark theme mode
+                            color: themeState.isDarkMode
+                                ? ConstantsDarkMode.indigoColor
+                                : ConstantsLightMode.indigoColor,
                             size: Constants.size15,
                           ),
                           Text(
                             Constants.faqsShortTxt,
                             style: TextStyle(
                               fontWeight: FontWeight.bold,
-                              color: Constants.indigoColor,
+
+                              /// Check the current light/dark theme mode
+                              color: themeState.isDarkMode
+                                  ? ConstantsDarkMode.indigoColor
+                                  : ConstantsLightMode.indigoColor,
                               fontSize: Constants.size15,
                             ),
                           )